+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 unreleased
 ==========
+* feat: Change view title and icon tooltips name changed
 * feat: Added compliance number field
 
 1.0.0 (2022-04-25)